new endpoint_cluster implementation
