+++ src/cpp/aws_http.cpp
@@ -59,29 +59,8 @@ public:
     if (epc != m.end())
       return epc->second;
 
-    auto lookup = [uri, key]() -> std::unique_ptr<std::pair<int, std::vector<std::pair<int, sockaddr_in6>>>> {
-#if ANON_LOG_NET_TRAFFIC > 1
-      anon_log("looking up sockaddr for " << key);
-#endif
-      auto dnsl = dns_lookup::get_addrinfo(uri.GetAuthority().c_str(), uri.GetPort());
-
-      std::unique_ptr<std::pair<int, std::vector<std::pair<int, sockaddr_in6>>>> ret(new std::pair<int, std::vector<std::pair<int, sockaddr_in6>>>());
-      ret->first = dnsl.first;
-      if (dnsl.first == 0)
-      {
-        for (auto a : dnsl.second)
-        {
-          a.sin6_port = htons(uri.GetPort());
-#if ANON_LOG_NET_TRAFFIC > 1
-          anon_log(" found: " << a);
-#endif
-          ret->second.push_back(std::make_pair(0 /*preference*/, a));
-        }
-      }
-      return std::move(ret);
-    };
-
-    return m[key] = endpoint_cluster::create(lookup, uri.GetScheme() == Scheme::HTTPS, uri.GetAuthority().c_str(), _tls.get());
+    return m[key] = endpoint_cluster::create(uri.GetAuthority().c_str(), uri.GetPort(),
+                                             uri.GetScheme() == Scheme::HTTPS, _tls.get());
   }
 
   std::shared_ptr<HttpResponse> MakeRequest(HttpRequest &request,
