+++ src/cpp/epc.cpp
@@ -1,5 +1,5 @@
 /*
- Copyright (c) 2015 Anon authors, see AUTHORS file.
+ Copyright (c) 2019 Anon authors, see AUTHORS file.
  
  Permission is hereby granted, free of charge, to any person obtaining a copy
  of this software and associated documentation files (the "Software"), to deal
@@ -20,562 +20,268 @@
  THE SOFTWARE.
 */
 
-#include "epc.h"
-#include <netdb.h>
+#include <epc.h>
+#include "dns_lookup.h"
+
+endpoint_cluster::endpoint_cluster(const char *host, int port,
+                                   bool do_tls,
+                                   const tls_context *tls_ctx,
+                                   int max_conn_per_ep,
+                                   int lookup_frequency_in_seconds)
+    : host_(host),
+      port_(port),
+      do_tls_(do_tls),
+      tls_ctx_(tls_ctx),
+      max_conn_per_ep_(max_conn_per_ep),
+      lookup_frequency_in_seconds_(lookup_frequency_in_seconds),
+      last_lookup_time_((struct timespec){}),
+      round_robin_index_(0),
+      looking_up_endpoints_(false)
+{
+}
 
 void endpoint_cluster::update_endpoints()
 {
+  auto addrs = dns_lookup::get_addrinfo(host_.c_str(), port_);
+
+  fiber_lock l(mtx_);
+  if (addrs.first != 0)
   {
-    fiber_lock lock(mtx_);
-    if (shutting_down_)
+    // if there are no current endpoints that are already known
+    // for this host name, then this error object will be thrown
+    // in the fiber has initiated this call to update_endpoints.
+    // In that case, the fact that it is an instance of fiber_io_error
+    // will cause the exponential backoff logic to kick in and
+    // re-attempt the operation - optimistically hoping that whatever
+    // is causing the error will resolve itself
+    lookup_err_ = std::unique_ptr<fiber_io_error>(new fiber_io_error(Log::fmt(
+        [&](std::ostream &msg) { msg << "dns lookup failed for: " << host_ << ", error: " << error_string(addrs.first); })));
+  }
+  else
+  {
+    // build a map which is the union of all currently
+    // known endpoints plus the ones that were just
+    // returned in the call to get_adddrinfo.
+    auto now = cur_time();
+    std::map<sockaddr_in6, std::shared_ptr<endpoint>> endpoints;
+    for (auto &ep : endpoints_)
+      endpoints.insert(std::make_pair(ep->addr_, ep));
+    for (auto &addr : addrs.second)
+    {
+      auto it = endpoints.find(addr);
+      if (it == endpoints.end())
+        endpoints.insert(std::make_pair(addr, std::make_shared<endpoint>(addr)));
+      else
+        it->second->last_lookup_time_ = now;
+    }
+
+    // we use a simple policy rule that we are willing to
+    // keep attempting to use any ip address that was returned
+    // by getaddrinfo up to 10 times whatever the
+    // lookup_frequency_in_seconds_ is.  Any time we throw
+    // an exception due to either failure to connect or
+    // call an endpoint, if the socket in question is the
+    // only one for that endpoint, we delete the endpoint.
+    // So this will end up deleting endpoints that truly
+    // become invalid after lookup_frequency_in_seconds_
+    // and before lookup_frequency_in_seconds_ * 10
+    auto oldest = now - lookup_frequency_in_seconds_ * 10;
+    auto it = endpoints.begin();
+    while (it != endpoints.end())
+    {
+      auto next = it;
+      next++;
+      if (it->second->last_lookup_time_ < oldest)
+      {
+#if ANON_LOG_NET_TRAFFIC > 1
+        anon_log("aging out endpoint " << it->second->addr_ << ", because it was " << to_seconds(now - it->second->last_lookup_time_) << " seconds old");
+#endif
+        endpoints.erase(it);
+      }
+      it = next;
+    }
+    endpoints_.resize(endpoints.size());
+    int indx = 0;
+    for (auto &p : endpoints)
+    {
+#if ANON_LOG_NET_TRAFFIC > 1
+      anon_log(" using " << p.second->addr_ << ", lookup age: " << to_seconds(now - p.second->last_lookup_time_) << " seconds");
+#endif
+      endpoints_[indx++] = p.second;
+    }
+  }
+
+  last_lookup_time_ = cur_time();
+  looking_up_endpoints_ = false;
+  cond_.notify_all();
+}
+
+// This is called when there is some error that has occurred
+// on a socket related to this endpoint.  If it turns out that
+// that socket was the only one in existence for this endpoint
+// then remove the matching entry in the endpoints_ vector.
+// This causes that endpoint to get deleted (when the other
+// shared pointers to it are destructed) and means that the
+// endpoint_cluster itself will not attempt to use this endpoint
+// again without first going back through dns resolution.
+// This cleans up cases where the ip address itself is no
+// longer good and dns lookup is correctly no longer reporting
+// that ip address.  We have a policy of continuing to attempt
+// to use that ip address for some period after dns lookup has
+// returned it.
+void endpoint_cluster::erase_if_empty(const std::shared_ptr<endpoint> &ep)
+{
+#if ANON_LOG_NET_TRAFFIC > 1
+  anon_log("endpoint_cluster::erase_if_empty");
+#endif
+  {
+    fiber_lock l(ep->mtx_);
+    if (ep->socks_.size() != 0 || ep->outstanding_requests_ != 1)
+    {
+#if ANON_LOG_NET_TRAFFIC > 1
+      anon_log("endpoint_cluster::erase_if_empty, not emptying, ep->socks_.size(): " << ep->socks_.size() << ", ep->outstanding_requests_: " << ep->outstanding_requests_);
+#endif
       return;
-    update_running_ = true;
+    }
   }
 
-  // call the lookup function to get
-  // the list of ip addrs
-  auto endpoints = std::move(lookup_());
-
-  fiber_lock lock(mtx_);
-
-  int err = endpoints->first;
-
-  if (err != 0 || endpoints->second.empty())
+  fiber_lock l(mtx_);
+  auto it = endpoints_.begin();
+  while (it != endpoints_.end())
   {
-    // if this lookup failed (for example there was a problem reaching dns),
-    // but we previously were able to look things up and so have a non-empty
-    // endpoint_ list, then we ignore this error and continue to use the
-    // existing list.  But if we currently have no endpoints, for example
-    // if this is the first time we are calling lookup, and it fails, then
-    // we set this epc into an error state, allowing any fibers waiting in
-    // with_connected_pipe to see the error and exit.
-    if (endpoints_.empty())
+    if (*it == ep)
     {
-      lookup_error_ = err;
-      connections_possible_cond_.notify_all();
-    }
-  }
-  else
-  {
-
-    if (endpoints_.size() == 0) {
-
-    // in case we were previously in an error state
-    lookup_error_ = 0;
-
-    // for less typing...
-    const auto &eps = endpoints->second;
-
-    total_possible_requests_ = eps.size() * max_conn_per_ep_;
-
-    // move all endpoints out of the existing endpoints_ map and
-    // into a temporary map that is sorted by ip_addr
-    std::map<struct sockaddr_in6, std::unique_ptr<endpoint>> tmp_map;
-    for (auto it = endpoints_.begin(); it != endpoints_.end(); ++it)
-      tmp_map[it->second->addr_] = std::move(it->second);
-
-    // now empty or existing map;
-    endpoints_.clear();
-
-    // walk through all of the pref/addrs returned by lookup
-    int cur_outstanding_requests = 0;
-    for (auto ep : eps)
-    {
-
-      // is this an existing endpoint?
-      auto existing = tmp_map.find(ep.second);
-      if (existing != tmp_map.end())
-      {
-
-        // yes, move it back into endpoints_, but use the (possibly updated)
-        // preference reported by this call to lookup
-        cur_outstanding_requests += existing->second->outstanding_requests_;
-        existing->second->is_detached_ = false;
-        endpoints_.insert(std::make_pair(ep.first, std::move(existing->second)));
-      }
-      else
-      {
-
-        // no, make a new endpoint and put that in endpoints_
-        endpoints_.insert(std::make_pair(ep.first, std::unique_ptr<endpoint>(new endpoint(ep.second))));
-      }
-    }
-
-    // now handle any endpoints that have become "detached" -
-    // that is, they were in endpoints_ at the start of this function
-    // but were not returned by lookup (it no longer considers them
-    // as valid ip_addrs), so they are now still valid pointers
-    // in tmp_map.  If there are outstanding requests on that endpoint
-    // we keep it in the endpoints_ map until those requests have completed.
-    // otherwise we let the tmp_map dtor delete them.
-    for (auto it = tmp_map.begin(); it != tmp_map.end(); ++it)
-    {
-      if (it->second)
-      {
-        if (it->second->outstanding_requests_ != 0)
-        {
-          it->second->is_detached_ = true;
-          endpoints_.insert(std::make_pair(std::numeric_limits<int>::max(), std::move(it->second)));
-        }
-      }
-    }
-
-    cur_avail_requests_ = total_possible_requests_ - cur_outstanding_requests;
-    if (cur_avail_requests_ > 0)
-      connections_possible_cond_.notify_all();
-
-    }
-  }
-
-  if (!shutting_down_ && lookup_frequency_seconds_ > 0)
-  {
-    std::weak_ptr<endpoint_cluster> wp = shared_from_this();
-    update_task_ = io_dispatch::schedule_task(
-        [wp] {
-          auto stack_size = 16 * 1024 - 256;
-          fiber::run_in_fiber(
-              [wp] {
-                auto ths = wp.lock();
-                if (ths)
-                  ths->update_endpoints();
-              },
-              stack_size, "endpoint_cluster, update_endpoints");
-        },
-        cur_time() + lookup_frequency_seconds_);
-  }
-
-  update_running_ = false;
-  update_cond_.notify_one();
-}
-
-void endpoint_cluster::idle_socket_sweep()
-{
-  auto mx_time = cur_time() + 60;
-  for (auto it = endpoints_.begin(); it != endpoints_.end(); it++)
-  {
-    for (auto it2 = it->second->socks_.begin(); it2 != it->second->socks_.end();)
-    {
-      if (!(*it2)->in_use_ && (*it2)->last_used_time_ < mx_time)
-        it->second->socks_.erase(it2++);
-      else
-        it2++;
-    }
-  }
-}
-
-void endpoint_cluster::backoff(endpoint *ep, const struct timespec &start_time, int explicit_seconds)
-{
-  auto now = cur_time();
-  if (explicit_seconds > 0)
-  {
-#if ANON_LOG_NET_TRAFFIC > 0
-    anon_log("backing off of: " << ep->addr_ << ", for the next " << explicit_seconds << " (explicit) seconds");
+#if ANON_LOG_NET_TRAFFIC > 1
+      anon_log("endpoint_cluster::erase_if_empty emptying");
 #endif
-    ep->next_avail_time_ = now + explicit_seconds;
-  }
-  else
-  {
-    auto error_time = now - start_time;
-
-    // for each time we get an error trying to connect or talk to this ip_addr,
-    // double the amount of time we will wait before attempting to connect
-    // to it again.
-    auto btime = (1 << (ep->backoff_exp_++)) * error_time;
-    if (ep->backoff_exp_ > 20)
-      ep->backoff_exp_ = 20;
-
-    if (btime.tv_sec > 120)
-    {
-      btime.tv_sec = 120;
-      btime.tv_nsec = 0;
+      endpoints_.erase(it);
+      return;
     }
-
-#if ANON_LOG_NET_TRAFFIC > 0
-    anon_log("backing off of: " << ep->addr_ << ", for the next " << to_seconds(btime) << " seconds");
-#endif
-
-    // TODO: fix this...
-    cur_avail_requests_ = 0;
-
-    ep->next_avail_time_ = now + btime;
+    it++;
   }
-
-  // schedule a task to call connections_possible_cond_.notify_all()
-  // when we have allowed enough time to pass for it to be worth
-  // trying this ip_addr again.
-  //
-  // note that fiber_cond::notify_all can only be called in a fiber,
-  // and the task mechanism in io_dispatch does not call our functor
-  // in a fiber.  So here the functor creates a fiber and executes
-  // notify_all in there.  It is the fiber exiting condition that
-  // actually schedules the notified fibers to run.
-  std::weak_ptr<endpoint_cluster> wp = shared_from_this();
-  auto wakeup_task = io_dispatch::schedule_task(
-      [wp] {
-        fiber::run_in_fiber(
-            [wp] {
-              auto ths = wp.lock();
-              if (ths)
-              {
-                fiber_lock lock(ths->mtx_);
-
-                // TODO: fix this
-                int cur_outstanding_requests = 0;
-                for (auto ep = ths->endpoints_.begin(); ep != ths->endpoints_.end(); ep++)
-                {
-                  if (!ep->second->is_detached_)
-                    cur_outstanding_requests += ep->second->outstanding_requests_;
-                }
-
-                ths->cur_avail_requests_ = ths->total_possible_requests_ - cur_outstanding_requests;
-                if (ths->cur_avail_requests_ > 0)
-                  ths->connections_possible_cond_.notify_all();
-
-                auto ct = cur_time();
-                auto it = ths->io_retry_tasks_.begin();
-                while (it != ths->io_retry_tasks_.end() && it->first <= ct)
-                {
-                  io_dispatch::remove_task(it->second);
-                  ths->io_retry_tasks_.erase(it++);
-                }
-              }
-            },
-            fiber::k_default_stack_size, "endpoint_cluster::backoff");
-      },
-      ep->next_avail_time_);
-  io_retry_tasks_.insert(std::make_pair(ep->next_avail_time_, wakeup_task));
-
-  // this exception gets caught in with_connected_pipe and causes that
-  // to call us (do_with_connected_pipe) again immediately.  But we have
-  // set this ip_addr's next_avail_time_ out some amount, meaning
-  // that if this epc has other available ip_addrs those will be tried.
-  // If no other ip_addrs are available then the function will wait
-  // in connections_possible_cond_.wait (above in this routine) until
-  // one of the io_dispatch scheduled tasks (immediately above) notifies
-  // of the possibility of reconnecting.  If this epc has only one ip_addr
-  // this will execute a fairly traditional "exponential backoff" design.
-  throw backoff_error();
 }
 
-// a bit complicated because of all of the error/retry things it deals with...
 void endpoint_cluster::do_with_connected_pipe(const std::function<void(const pipe_t *pipe)> &f)
 {
-
-  // this outer loop performs reconnects to an existing,
-  // working ip_addr when a once-functional socket is now
-  // throwing io errors on read and write
-  while (true)
+  // if there are currently no available endpoints, or if it has been too long since we have
+  // last looked up endpoints, then restart the lookup process.  If there are no endpoints
+  // then wait until the lookup is complete, otherwise continue to use the endpoints we
+  // already know about and let the lookup complete asynchronously.  If there was an error
+  // attempting the lookup throw that error if there are no current endpoints, otherwise
+  // ignore the error
+  std::shared_ptr<endpoint> ep;
+  std::weak_ptr<endpoint> wep;
   {
-
-    // when we exit the try_again loop
-    // we will have set these 2 to the
-    // socket/pipe thing we are trying
-    // to use when calling 'f'
-    endpoint *ep = 0;
-    endpoint::sock *epsock = 0;
-    struct timespec ct;
-
-    // this next loop deals with the inability to find
-    // a currently usable ip_addr associated with this
-    // epc.  Its important steps are to 1) lock mtx_,
-    // 2) wait until cur_avail_requests_ > 0, and
-    // then 3) try to find/connect to an ip_addr
-    bool try_again = true;
-    while (try_again)
+    fiber_lock l(mtx_);
+    if (endpoints_.size() == 0 || to_seconds(cur_time() - last_lookup_time_) > lookup_frequency_in_seconds_)
     {
-      fiber_lock lock(mtx_);
-
-      // sleep until it is worth trying a connection
-      while (cur_avail_requests_ == 0 && lookup_error_ == 0)
-        connections_possible_cond_.wait(lock);
-
-      // if the lookup has failed, then throw.  This
-      // transfers the error seen in the update_endpoints
-      // fiber to this one.
-      if (lookup_error_ != 0)
+      if (!looking_up_endpoints_)
       {
-        if (lookup_error_ > 0)
-        {
-          errno = lookup_error_;
-          do_error("endpoint_cluster lookup");
-        }
-        else
-        {
-          anon_throw(std::runtime_error, "endpoint_cluster lookup failed: " << gai_strerror(lookup_error_));
-        }
+        looking_up_endpoints_ = true;
+        lookup_err_.reset();
+        std::weak_ptr<endpoint_cluster> wp = shared_from_this();
+        fiber::run_in_fiber([wp] {
+          auto ths = wp.lock();
+          if (ths)
+          {
+            ths->update_endpoints();
+          }
+        });
       }
-
-      // if we are shutting down then we abort these
-      // calls
-      if (shutting_down_)
-        throw std::runtime_error("endpoint_cluster destructed while fiber waiting for available endpoints, aborting call to with_connected_pipe");
-
-      // backoff etc will mark certain ip_addrs as
-      // temporarily unavailable using a timestamp,
-      // so get the current time so we can compare
-      ct = cur_time();
-
-      // figure out which endpoint/ip_addr we are going to use.
-      // the 'endpoints_' list is in order, sorted by preference,
-      // so we will use the first one that is currently usable
-      for (auto ep_it = endpoints_.begin(); ep_it != endpoints_.end(); ep_it++)
+      while (endpoints_.size() == 0)
       {
+        cond_.wait(l);
+        if (lookup_err_)
+          throw *lookup_err_;
+      }
+    }
+    ep = endpoints_[round_robin_index_++ % endpoints_.size()];
+    wep = ep;
+  }
 
-        // just for less typing...
-        ep = ep_it->second.get();
+  std::shared_ptr<endpoint::sock> sock;
+  {
+    fiber_lock l(ep->mtx_);
 
-        // we can only use this ep if:
-        //  1) it isn't already at the maximum number of allowed outstanding requests
-        //  2) it isn't in a "detached" state, and
-        //  3) we haven't marked it as temporarily unavailable
-        if (ep->outstanding_requests_ < max_conn_per_ep_ && !ep->is_detached_ && ep->next_avail_time_ < ct)
-        {
+    while (ep->outstanding_requests_ >= max_conn_per_ep_)
+      ep->cond_.wait(l);
+    ++ep->outstanding_requests_;
 
-          // Round Robin
-          // how many ip_addrs have this same preference?
-          auto ep2_it = ep_it;
-          ++ep2_it;
-          unsigned int num_with_same_pref = 1;
-          while ((ep2_it != endpoints_.end()) && (ep2_it->first == ep_it->first))
-          {
-
-            // ignoring any that we can't use at the moment
-            if (!ep2_it->second->is_detached_ && (ep2_it->second->outstanding_requests_ < max_conn_per_ep_) && (ep2_it->second->next_avail_time_ < ct))
-              ++num_with_same_pref;
-
-            ++ep2_it;
-          }
-          // if more than one, reset ep to the next round-robin ip_addr
-          if (num_with_same_pref > 1)
-          {
-            unsigned int which = round_robin_index_++ % num_with_same_pref;
-            for (unsigned int i = 0; i < which; i++)
-            {
-              ++ep_it;
-
-              // remember to skip the ones we ignored above when we were counting
-              while (!(!ep_it->second->is_detached_ && (ep_it->second->outstanding_requests_ < max_conn_per_ep_) && (ep_it->second->next_avail_time_ < ct)))
-                ++ep_it;
-            }
-            // reset ep to the one we are actually going to use
-            ep = ep_it->second.get();
-          }
-
-          // we're going to make a request to this endpoint
-          ++ep->outstanding_requests_;
-
-          // which will decrement the total number of available requests
-          --cur_avail_requests_;
-
-          // is there already an existing, unused socket connected to this ip_addr?
-          std::list<std::unique_ptr<endpoint::sock>>::iterator s_it;
-          for (s_it = ep->socks_.begin(); s_it != ep->socks_.end(); s_it++)
-          {
-            if (!(*s_it)->in_use_)
-              break;
-          }
-
-          // if no existing, unused sockets then connect a new one
-          if (s_it == ep->socks_.end())
-          {
-            // sleep this fiber until the connection attempt completes.  The
-            // return value is a pair<int err_code, unique_ptr<fiber_pipe>>
-            //
-            // note that we do a manual unlock followed by re-lock because we
-            // don't want the mutex locked over this call to connect.  But we
-            // want to run in a single stack frame so that the exception handling
-            // is simpler.  Alternate designs would be to run the connect
-            // completion code in a fiber, but then we would need to replicate
-            // some exception handling logic, and it would be harder to do there.
-            // We are being somewhat careful here to mark all data structures
-            // we are referencing as being "used" prior to unlocking, and this
-            // ensures that they won't be deleted or reused before we get back
-            // and lock again.  And because connect can throw (for example, if
-            // we run out of file descriptors or something similar) we ensure
-            // we remove the in_use settings and then re-lock before exiting
-            // if that were to happen.
-            mtx_.unlock();
-            std::pair<int, std::unique_ptr<fiber_pipe>> con;
-
-            // TODO, start a timer to detect "suspiciously slow" connection attempts.
-            // probably some research into what linux actually does with unanswered
-            // SYN segments, and the frequency of resends for them.  Whatever happens
-            // here should compliment that linux behavior, not simply reimplement it.
-
-            try
-            {
-              con = tcp_client::connect((struct sockaddr *)&ep->addr_, ep->addr_.sin6_family == AF_INET6 ? sizeof(struct sockaddr_in6) : sizeof(struct sockaddr_in));
-#if ANON_LOG_NET_TRAFFIC > 2
-              anon_log("epc connected to: " << ep->addr_);
+    if (ep->socks_.size() != 0)
+    {
+      sock = ep->socks_.front();
+      ep->socks_.pop();
+#if ANON_LOG_NET_TRAFFIC > 1
+      anon_log("epc reused connection to " << ep->addr_);
 #endif
-            }
-            catch (...)
-            {
-              mtx_.lock();
-              ep_exit(ep);
-              throw;
-            }
-            mtx_.lock();
-
-            if (shutting_down_)
-            {
-              ep_exit(ep);
-              throw std::runtime_error("endpoint_cluster destructed while fiber waiting on tcp_client::connect, aborting call to with_connected_pipe");
-            }
-
-            switch (con.first)
-            { // con.first is the error code part of the return value
-
-            case 0:
-            { // connected, we're good to go!
-              std::unique_ptr<pipe_t> pipe;
-              if (do_tls_)
-                pipe = std::unique_ptr<pipe_t>(new tls_pipe(std::move(con.second),
-                                                            true,                     // client (not server)
-                                                            host_name_for_tls_ != "", // verify_peer
-                                                            host_name_for_tls_ != "", // doSNI
-                                                            host_name_for_tls_.c_str(),
-                                                            *tls_ctx_));
-              else
-                pipe = std::unique_ptr<pipe_t>(con.second.release());
-              ep->socks_.push_back(std::unique_ptr<endpoint::sock>(new endpoint::sock(std::move(pipe))));
-              s_it = ep->socks_.end();
-              --s_it;
-              epsock = s_it->get();
-            }
-            break;
-
-              //  Here is a basic description of the three interesting, similar
-              //  errors that we care about that can come back from connect.
-              //
-              //  ENETUNREACH:  no SYN segment was answered by the server, and at least
-              //                one such segment generated an ICMP "destination unreachable"
-              //                response in some intermediate router (note that connect will send
-              //                more than one SYN if it doesn't hear back).  Allocated, but currently
-              //                unassigned ip addresses can cause this.  These are more common in
-              //                ipv6 than ipv4.
-              //
-              //  ETIMEDOUT:    no SYN segment was answered by the server or intermediate routers.
-              //                We received no response whatsoever to any of our SYN segments.
-              //
-              //  ECONNREFUSED: server responded to our SYN with a RST.  Generally indicates that
-              //                the server machine/os is running and reachable, but that no server
-              //                process is listening on the port number we are requesting.  Intermediate
-              //                load balancers and proxies greatly complicate this picture, because
-              //                in those cases it is the intermediate machine that is acting like
-              //                the server from the perspective of these network errors.
-              //
-              //  Particularly because of the last point in ECONNREFUSED, where proxies complicate the
-              //  issues of understanding what each error value actually implies about the state of the
-              //  network and servers on it, we don't attempt to treat any of these errors differently
-              //  from each other.  We simply adopt a basic strategy of backing off of our attempt to
-              //  connect to that endpoint.  We take into account the length of time it takes to get the
-              //  error back in such a way that we more quickly back off of ip_addrs that return errors
-              //  more slowly.
-
-            case ENETUNREACH:
-            case ECONNREFUSED:
-            case ETIMEDOUT:
-              ep_exit(ep);
-              backoff(ep, ct); // ends with a "throw backoff_error();"
-
-              // current decision is to not special-case this error code.  It would represent
-              // something like a very poorly configured machine, or perhaps some kind of resource
-              // leak in our logic.  Neither of these are things we can easily account for here.
-              //case EAGAIN:        // we ran out of available (local) port numbers, so couldn't do the "bind" step
-
-            default: // some unknown problem
-              ep_exit(ep);
-              if (con.first)
-              {
-                errno = con.first;
-                do_error("tcp_client::connect(" << ep->addr_ << ")");
-              }
-              else
-              {
-                anon_throw(std::runtime_error, "tcp_client::connect(" << ep->addr_ << "): " << gai_strerror(con.first));
-              }
-            }
-          }
-          else
-          { // end of "if (s_it == ep->end())"
-
-            // here we found an unused, already-connected socket, so use it.
-            epsock = s_it->get();
-            epsock->in_use_ = true;
-
-          } // end of else of "if (s_it == ep->end())"
-
-          // ep, and epsock are now set
-
-          try_again = false; // so we exit the "while (try_again)" loop after we...
-          break;             // break out of the "for (auto ep_it = endpoints_.begin(); ep_it != endpoints_.end(); ep_it++)" loop
-
-        } // end of "if (ep->outstanding_requests_ < max_conn_per_ep_)"
-
-      } // end of "for (auto ep_it = endpoints_.begin(); ep_it != endpoints_.end(); ep_it++)"
-
-    } // end of "while (try_again)"
-
-    // mtx_ is now unlocked
-    // ep and epsock are set pointing to the ip_addr/socket/pipe_t we are supposed to try.
-    // if a read/write exception (fiber_io_error) goes off at this point we immediately
-    // close our side of the socket, deleting the corresponding 'sock' object and retry
-    // this function.  By default this condition does _not_ cause us to back off of trying
-    // this ip_addr.  However, fiber_io_error has a public bool "backoff" which, if set,
-    // will cause this us to backoff.  The idea is that code which might do something like
-    // parse an HTTP response, looking for a 503, could throw a specially constructed
-    // fiber_io_error to trigger this behavior.  But normally, a simple error in attempting
-    // to read from or write to a once-valid socket will just assume that the other side
-    // closed its end, and so we are free to immediately attempt to reconnect.
-
-    try
-    {
-      epsock->last_used_time_ = cur_time();
-      f(epsock->pipe_.get());
-      fiber_lock lock(mtx_);
-      ++ep->success_count_;
-      ep->backoff_exp_ = 0;
-      ep_exit(ep);
-      epsock->in_use_ = false;
-      return;
     }
-    catch (const fiber_io_error &e)
+    else
     {
-      fiber_lock lock(mtx_);
-      ep_exit(ep);
-
-      // note that we treat the case where the server
-      // closes a connection immediately after accepting
-      // (so we get a fiber_io_error as soon as we try
-      // to use the pipe the first time) as a form of
-      // backoff request from the server
-
-      if (!ep->success_count_ || !e.backoff_ || e.close_socket_hint_)
+      l.unlock();
+      auto conn = tcp_client::connect((struct sockaddr *)&ep->addr_, ep->addr_.sin6_family == AF_INET6 ? sizeof(struct sockaddr_in6) : sizeof(struct sockaddr_in));
+      if (conn.first != 0)
       {
-        for (auto s_it = ep->socks_.begin(); s_it != ep->socks_.end(); s_it++)
-        {
-          if (s_it->get() == epsock)
-          {
-            ep->socks_.erase(s_it);
-            break;
-          }
-        }
+        erase_if_empty(ep);
+        anon_throw(fiber_io_error, "tcp connect failed for " << ep->addr_ << ", error: " << error_string(conn.first));
       }
-      else
-        epsock->in_use_ = false;
-      if (!ep->success_count_)
-        backoff(ep, ct);
-      if (e.backoff_)
-        backoff(ep, timespec(), e.backoff_seconds_);
-    }
-    catch (...)
-    {
-      fiber_lock lock(mtx_);
-      ep_exit(ep);
-      epsock->in_use_ = false;
-      throw;
-    }
 
-  } // end of "while (true)", but we do the normal return when the 'try' clause above does not throw
+      std::unique_ptr<pipe_t> pipe;
+      if (do_tls_)
+        pipe = std::unique_ptr<pipe_t>(new tls_pipe(std::move(conn.second),
+                                                    true, // client (not server)
+                                                    true, // verify_peer
+                                                    true, // doSNI
+                                                    host_.c_str(),
+                                                    *tls_ctx_));
+      else
+        pipe = std::unique_ptr<pipe_t>(conn.second.release());
+      sock = std::shared_ptr<endpoint::sock>(new endpoint::sock(std::move(pipe)));
+#if ANON_LOG_NET_TRAFFIC > 1
+      anon_log("epc established new connection to " << ep->addr_);
+#endif
+    }
+  }
+
+  // we let go of the endpoint itself and only hole
+  // the weak pointer to it across the call.  This
+  // lets it timeout and get deleted more smoothly.
+  ep.reset();
+
+  try
+  {
+    // call the function with the connected pipe
+    f(sock->pipe_.get());
+
+    // if that function did not throw an exception
+    // cache the socket/pipe for next time, but since
+    // it is also possible that the endpoint itself was
+    // deleted, we check for that.  If the end point was
+    // deleted we just let the socket/pipe close/delete
+    ep = wep.lock();
+    if (ep)
+    {
+      fiber_lock l(ep->mtx_);
+      --ep->outstanding_requests_;
+      ep->socks_.push(sock);
+      ep->cond_.notify_all();
+    }
+    else
+      anon_log("epc, appears that endpoint was deleted prior to callback returning");
+  }
+  catch (...)
+  {
+    // if an exception was thrown we still need to
+    // decrement the number of outstanding_requests_
+    anon_log("exception thrown when executing with_connected_pipe call");
+    ep = wep.lock();
+    if (ep)
+    {
+      erase_if_empty(ep);
+      --ep->outstanding_requests_;
+      ep->cond_.notify_all();
+    }
+    else
+      anon_log("epc, appears that endpoint was deleted prior to callback throwing");
+    throw;
+  }
 }
